Fix proper defaulting/required of 'state' for gem module
The docs for 'gem' say state is required, but the actual code says it's not required.

If it's not included (and there's no default) then the if block falls through with no changes and no errors (it neither adds nor removes the gem).

This change synchronizes the docs with the code.  It also adds the sane default of state=present, which is consistent with all other modules with the 'state' field.